% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firedam.R
\docType{data}
\name{firedam}
\alias{firedam}
\title{Distance and damage from fire}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance of fire}
  \item{DAMAGE}{damage of fire}
  ...
}
}
\source{
\url{MATH4753 files}
}
\usage{
firedam
}
\description{
A dataset containing 15 samples of the distance and damage of fire
}
\keyword{datasets}
