% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mybin}
\alias{mybin}
\title{my binomial function}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of samples taken each iteration, with replacement}

\item{p}{probability of TRUE sample}
}
\value{
a bar graph showing the distribution and associated table
}
\description{
finds the number of successes in each iteration given a probability.
The function adds the number of iterations that received each number of successes,
divides by overall number of iterations, and represents the distribution as a percentage.
}
\examples{
mybin()


}
