% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrapping function for basic stats}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{vector to evaluate}

\item{fun}{function to apply}

\item{alpha}{confidence value}

\item{cx}{numeric character expansion factor}

\item{...}{other params for hist}
}
\value{
a list of confidence interval, function, vector, and results per iter
}
\description{

}
\examples{
x = 1:10
myboot2(x=x)
}
