% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mycltp}
\alias{mycltp}
\title{central limit theorem applied to poisson distribution}
\usage{
mycltp(n, iter, lambda = 10, ...)
}
\arguments{
\item{n}{sample size}

\item{iter}{number of iterations}

\item{lambda}{mean and sd of the poisson distribution}

\item{...}{other parameters of the main histogram}
}
\value{
returns a list of the means of each iteration and plots of probability density
for the mean, the poisson distribution, and a relative frequency plot
}
\description{
applies the central limit theorem to the poisson distribution
and displays a histogram of the means distribution density
}
\examples{
mycltp(5, 100)

}
