% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myddt}
\alias{myddt}
\title{my ddt function for Project 1}
\usage{
myddt(df, spec)
}
\arguments{
\item{df}{the ddt dataframe or one similar}

\item{spec}{the species selected by the user}
}
\value{
a named list with a plot of the subsetted df, df before and after
subsetting, and a relative frequency of river table
}
\description{
subsets the ddt df by a species given as a parameter, generates
a corresponding plot, df before and after subsetting, and a relative frequency
of river table
}
