% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myhyper}
\alias{myhyper}
\title{my hypergeometric function}
\usage{
myhyper(iter = 100, N = 20, r = 12, n = 5)
}
\arguments{
\item{iter}{number of iterations}

\item{N}{population size}

\item{r}{number of successes in population}

\item{n}{number of samples each iteration without replacement}
}
\value{
a bar plot of the distribution and associated table
}
\description{
a function which shows the hypergeometric distribution
of a sample. Shows probability of number of successes while sampling without replacement
}
\examples{
myhyper()

}
