% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mymlnorm}
\alias{mymlnorm}
\title{my ml for normal distribution}
\usage{
mymlnorm(x, mu, sig, ...)
}
\arguments{
\item{x}{data to evaluate}

\item{mu}{vector of potential mean values}

\item{sig}{vector of potential sd values}

\item{...}{}
}
\value{
data, maximum likelihood, coords
}
\description{

}
\examples{
mymlnorm(x=c(10,12,13,15,12,11,10),mu=seq(10,14,length=1000),sig=seq(0.1,4,length=1000),lwd=2,labcex=1)

}
