% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mymult}
\alias{mymult}
\title{my multinomial function}
\usage{
mymult(iter = 100, n = 10, p = c(1, 1, 1, 1)/4)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{number of samples taken each iteration, with replacement}

\item{p}{array of probabilities for each outcome, add to 1}
}
\value{
a barplot showing the relative frequencies and a corresponding table
}
\description{
finds the number of successes in each iteration given an array of probabilities, 1 per option.
The function adds the number of iterations that received each number of successes,
divides by overall number of iterations, and represents the distribution as a percentage.
}
\examples{
mymult()

}
