% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myncurve}
\alias{myncurve}
\title{my probability curve function}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{average of distribution}

\item{sigma}{sd of distribution}

\item{a}{upper limit for probability, filled region}
}
\value{
plot of norm distribution w/ filled region, probability as list
}
\description{
uses mean and sd to create normal distribution, calculates
probability below a and shows as shaded region and prints value to console
}
\examples{
myncurve(5, 2, 3)

}
