% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{myread}
\alias{myread}
\title{my read function}
\usage{
myread(dird, csv)
}
\arguments{
\item{dird}{directory csv is stored in}

\item{csv}{the title of the .csv title as a string}
}
\value{
table containing data
}
\description{
This function reads .csv data into the environment
}
\examples{
\dontrun{dird = "C:\\\\Users\\\\Admin\\\\Desktop\\\\Applied Statistical Methods\\\\Lab 4\\\\"
data = myread(dird, 'SPRUCE.csv')}

}
