% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mysample}
\alias{mysample}
\title{my sample function}
\usage{
mysample(n, N = 10, iter = 10, time = 0.1, last = FALSE, rep = TRUE)
}
\arguments{
\item{n}{number of samples each iteration}

\item{N}{size of each sample}

\item{iter}{number of iterations}

\item{time}{time between showing each plot}

\item{last}{indicates that only last bar plot should be displayed}

\item{rep}{with or without replacement}
}
\value{
a barplot of the distribution and a corresponding table
}
\description{
takes n samples each iteration with replacement and
shows the resultant distribution
}
\examples{
mysample(5)

}
