% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spruce.R
\docType{data}
\name{spruce}
\alias{spruce}
\title{Heights and breast height diameters of spruce trees}
\format{
A data frame with 36 rows and 2 variables:
\describe{
  \item{Height}{height of spruce tree}
  \item{BHDiameter}{breast height diameter of spruce tree}
  ...
}
}
\source{
\url{MATH4753 files}
}
\usage{
spruce
}
\description{
A dataset with the heights and breast height diameters
of spruce trees which can be used to infer height from BHD
}
\keyword{datasets}
